# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SAMT Unified Backend API
  version: 1.0.0
  description: Generated by scripts/generate-openapi.mjs from service OpenAPI docs exposed via API Gateway.
servers:
  - url: http://localhost:9080
...
paths:
...
  /api/project-configs/{id}:
    get:
      tags:
        - project-config-controller
      summary: Get configuration by id
      operationId: getConfig
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Configuration retrieved
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/projectConfig_ApiResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
    put:
      tags:
        - project-config-controller
      operationId: updateConfig
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/projectConfig_UpdateConfigRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/projectConfig_ApiResponseConfigResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
    delete:
      tags:
        - project-config-controller
      operationId: deleteConfig
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/projectConfig_ApiResponseDeleteResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
  /api/project-configs:
    post:
      tags:
        - project-config-controller
      summary: Create configuration
      operationId: createConfig
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/projectConfig_CreateConfigRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/projectConfig_ConfigEnvelope'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
  /api/project-configs/{id}/verify:
    post:
      tags:
        - project-config-controller
      summary: Verify configuration
      operationId: verifyConfig
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/projectConfig_VerificationEnvelope'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          description: Verification service unavailable
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/projectConfig_ApiResponse'
  /api/project-configs/admin/{id}/restore:
    post:
      tags:
        - project-config-controller
      operationId: restoreConfig
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/projectConfig_ApiResponseRestoreResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
...
  /api/project-configs/group/{groupId}:
    get:
      tags:
        - project-config-controller
      operationId: getConfigByGroupId
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/projectConfig_ConfigEnvelope'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
...
components:
  schemas:
...
    projectConfig_UpdateConfigRequest:
      type: object
      properties:
        jiraHostUrl:
          type: string
          maxLength: 255
          minLength: 0
          pattern: https?://[a-zA-Z0-9.-]+\.(atlassian\.net|jira\.com)(/.*)?
        jiraEmail:
          type: string
          maxLength: 255
          minLength: 1
          format: email
        jiraApiToken:
          type: string
          maxLength: 500
          minLength: 100
          pattern: ^ATATT[A-Za-z0-9+/=_-]{95,495}$
        githubRepoUrl:
          type: string
          maxLength: 512
          minLength: 0
          pattern: https://github\.com/[\w-]+/[\w-]+
        githubToken:
          type: string
          maxLength: 255
          minLength: 40
          pattern: ^(ghp_[A-Za-z0-9]{36,}|github_pat_[A-Za-z0-9_]+)$
    projectConfig_ApiResponseConfigResponse:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/projectConfig_ConfigResponse'
        error:
          type: string
        message:
          type: string
        path:
          type: string
        correlationId:
          type: string
        degraded:
          type: boolean
    projectConfig_ConfigResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        groupId:
          type: integer
          format: int64
        jiraHostUrl:
          type: string
        jiraEmail:
          type: string
        jiraApiToken:
          type: string
        githubRepoUrl:
          type: string
        githubToken:
          type: string
        state:
          type: string
        lastVerifiedAt:
          type: string
          format: date-time
          nullable: true
        invalidReason:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    projectConfig_CreateConfigRequest:
      type: object
      properties:
        groupId:
          type: integer
          format: int64
          minimum: 1
        jiraHostUrl:
          type: string
          maxLength: 255
          minLength: 0
          pattern: https?://[a-zA-Z0-9.-]+\.(atlassian\.net|jira\.com)(/.*)?
        jiraEmail:
          type: string
          maxLength: 255
          minLength: 0
          format: email
        jiraApiToken:
          type: string
          maxLength: 500
          minLength: 100
          pattern: ^ATATT[A-Za-z0-9+/=_-]{95,495}$
        githubRepoUrl:
          type: string
          maxLength: 512
          minLength: 0
          pattern: https://github\.com/[\w-]+/[\w-]+
        githubToken:
          type: string
          maxLength: 255
          minLength: 40
          pattern: ^(ghp_[A-Za-z0-9]{36,}|github_pat_[A-Za-z0-9_]+)$
      required:
        - groupId
        - jiraHostUrl
        - jiraEmail
        - jiraApiToken
        - githubRepoUrl
        - githubToken
    projectConfig_ApiResponse:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
        message:
          type: string
        path:
          type: string
        correlationId:
          type: string
        degraded:
          type: boolean
    projectConfig_ApiResponseRestoreResponse:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/projectConfig_RestoreResponse'
        error:
          type: string
        message:
          type: string
        path:
          type: string
        correlationId:
          type: string
        degraded:
          type: boolean
    projectConfig_RestoreResponse:
      type: object
      properties:
        message:
          type: string
        configId:
          type: string
          format: uuid
        restoredAt:
          type: string
          format: date-time
        state:
          type: string
...
    projectConfig_ApiResponseDeleteResponse:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/projectConfig_DeleteResponse'
        error:
          type: string
        message:
          type: string
        path:
          type: string
        correlationId:
          type: string
        degraded:
          type: boolean
    projectConfig_DeleteResponse:
      type: object
      properties:
        message:
          type: string
        configId:
          type: string
          format: uuid
        deletedAt:
          type: string
          format: date-time
        retentionDays:
          type: integer
          format: int32
...
    StandardError:
      type: object
      properties:
        type:
          type: string
          example: https://api.example.com/errors/validation-error
        title:
          type: string
          example: Validation failed
        status:
          type: integer
          format: int32
          example: 400
        detail:
          type: string
          example: Invalid UUID
        instance:
          type: string
          example: /api/users/not-a-number
        timestamp:
          type: string
          format: date-time
...
    projectConfig_ConfigData:
      type: object
      properties:
        id:
          type: string
          format: uuid
        groupId:
          type: integer
          format: int64
        jiraHostUrl:
          type: string
        jiraEmail:
          type: string
          format: email
        jiraApiToken:
          type: string
        githubRepoUrl:
          type: string
        githubToken:
          type: string
        state:
          type: string
        lastVerifiedAt:
          type: string
          format: date-time
          nullable: true
        invalidReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    projectConfig_ConfigEnvelope:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/projectConfig_ConfigData'
        timestamp:
          type: string
          format: date-time
    projectConfig_VerificationEnvelope:
      type: object
      properties:
        data:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
...
  responses:
    Error400:
      description: 400 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error401:
      description: 401 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error403:
      description: 403 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error404:
      description: 404 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error405:
      description: 405 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error409:
      description: 409 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error415:
      description: 415 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error500:
      description: 500 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    identity_bearerAuth:
      type: http
      description: Enter JWT access token
      scheme: bearer
      bearerFormat: JWT
    userGroup_bearerAuth:
      type: http
      description: JWT token from identity-service
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT
    projectConfig_Bearer Authentication:
      type: http
      description: Enter your JWT token from Identity Service
      scheme: bearer
      bearerFormat: JWT
    sync_bearerAuth:
      type: http
      description: Paste the JWT only. Swagger UI adds the Bearer prefix automatically.
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT
    report_bearerAuth:
      type: http
      description: Paste the JWT only. Swagger UI adds the Bearer prefix automatically.
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT
...